budgets: Reorganizing app_users pages
